Fixing issue when app is deployed to root context - https://github.com/DependencyTrack/dependency-track/issues/665
