+++ src/i18n/locales/en.json
@@ -23,6 +23,7 @@
     "dashboard": "Dashboard",
     "projects": "Projects",
     "components": "Components",
+    "component_search": "Component Search",
     "component": "Component",
     "dependencies": "Dependencies",
     "vulnerabilities": "Vulnerabilities",
@@ -114,8 +115,8 @@
     "false_positive": "False Positive",
     "not_affected": "Not Affected",
     "updated": "Updated",
-    "add_dependency": "Add Dependency",
-    "remove_dependency": "Remove Dependency",
+    "add_component": "Add Component",
+    "remove_component": "Remove Component",
     "upload_bom": "Upload BOM",
     "upload": "Upload",
     "reset": "Reset",
@@ -163,12 +164,8 @@
     "component_hash_desc": "Specifies the observed hash value of the component using the preceding algorithm",
     "required_component_name": "The component name is required",
     "required_component_version": "The component version is required",
-    "component_updated": "Component updated",
-    "component_deleted": "Component deleted",
-    "component_created": "Component created",
-    "create_component": "Create Component",
-    "dependency_added": "Dependency added",
-    "dependency_removed": "Dependency removed",
+    "component_added": "Component added",
+    "component_removed": "Component removed",
     "profile": "Profile",
     "profile_update": "Update Profile",
     "profile_updated": "Profile updated",
