+++ src/shared/api.json
@@ -28,7 +28,6 @@
   "URL_LICENSE_CONCISE": "api/v1/license/concise",
   "URL_CWE": "api/v1/cwe",
   "URL_COMPONENT": "api/v1/component",
-  "URL_DEPENDENCY": "api/v1/dependency",
   "URL_VULNERABILITY": "api/v1/vulnerability",
   "URL_ANALYSIS": "api/v1/analysis",
   "URL_SEARCH": "api/v1/search",
