+++ src/views/portfolio/projects/ProjectComponents.vue
@@ -3,14 +3,14 @@
     <div id="toolbar">
       <div class="form-inline btn-spaced-group" role="form">
         <b-button size="md" variant="outline-primary"
-                  v-b-modal.projectAddDependencyModal
+                  v-b-modal.projectAddComponentModal
                   v-permission="PERMISSIONS.PORTFOLIO_MANAGEMENT">
-          <span class="fa fa-plus"></span> {{ $t('message.add_dependency') }}
+          <span class="fa fa-plus"></span> {{ $t('message.add_component') }}
         </b-button>
         <b-button size="md" variant="outline-primary"
                   @click="removeDependencies"
                   v-permission="PERMISSIONS.PORTFOLIO_MANAGEMENT">
-          <span class="fa fa-minus"></span> {{ $t('message.remove_dependency') }}
+          <span class="fa fa-minus"></span> {{ $t('message.remove_component') }}
         </b-button>
         <b-button size="md" variant="outline-primary"
                   v-b-modal.projectUploadBomModal
@@ -52,24 +52,24 @@
           },
           {
             title: this.$t('message.component'),
-            field: "component.name",
+            field: "name",
             sortable: true,
             formatter(value, row, index) {
-              let url = xssFilters.uriInUnQuotedAttr("../components/" + row.component.uuid);
+              let url = xssFilters.uriInUnQuotedAttr("../projects/" + row.project.uuid + "/" + row.uuid);
               return `<a href="${url}">${xssFilters.inHTMLData(value)}</a>`;
             }
           },
           {
             title: this.$t('message.version'),
-            field: "component.version",
+            field: "version",
             sortable: true,
             formatter(value, row, index) {
-              if (Object.prototype.hasOwnProperty.call(row.component, "repositoryMeta") && Object.prototype.hasOwnProperty.call(row.component.repositoryMeta, "latestVersion")) {
-                row.component.latestVersion = row.component.repositoryMeta.latestVersion;
-                if (row.component.repositoryMeta.latestVersion !== row.component.version) {
-                  return '<span style="float:right" data-toggle="tooltip" data-placement="bottom" title="Risk: Outdated component. Current version is: '+ xssFilters.inHTMLData(row.component.repositoryMeta.latestVersion) + '"><i class="fa fa-exclamation-triangle status-warning" aria-hidden="true"></i></span> ' + xssFilters.inHTMLData(row.component.version);
+              if (Object.prototype.hasOwnProperty.call(row, "repositoryMeta") && Object.prototype.hasOwnProperty.call(row.repositoryMeta, "latestVersion")) {
+                row.latestVersion = row.repositoryMeta.latestVersion;
+                if (row.repositoryMeta.latestVersion !== row.version) {
+                  return '<span style="float:right" data-toggle="tooltip" data-placement="bottom" title="Risk: Outdated component. Current version is: '+ xssFilters.inHTMLData(row.repositoryMeta.latestVersion) + '"><i class="fa fa-exclamation-triangle status-warning" aria-hidden="true"></i></span> ' + xssFilters.inHTMLData(row.version);
                 } else {
-                  return '<span style="float:right" data-toggle="tooltip" data-placement="bottom" title="Component version is the latest available from the configured repositories"><i class="fa fa-exclamation-triangle status-passed" aria-hidden="true"></i></span> ' + xssFilters.inHTMLData(row.component.version);
+                  return '<span style="float:right" data-toggle="tooltip" data-placement="bottom" title="Component version is the latest available from the configured repositories"><i class="fa fa-exclamation-triangle status-passed" aria-hidden="true"></i></span> ' + xssFilters.inHTMLData(row.version);
                 }
               } else {
                 return xssFilters.inHTMLData(common.valueWithDefault(value, ""));
@@ -78,7 +78,7 @@
           },
           {
             title: this.$t('message.latest_version'),
-            field: "component.latestVersion",
+            field: "latestVersion",
             sortable: false,
             visible: false,
             formatter(value, row, index) {
@@ -87,7 +87,7 @@
           },
           {
             title: this.$t('message.group'),
-            field: "component.group",
+            field: "group",
             sortable: true,
             formatter(value, row, index) {
               return xssFilters.inHTMLData(common.valueWithDefault(value, ""));
@@ -95,7 +95,7 @@
           },
           {
             title: this.$t('message.internal'),
-            field: "component.isInternal",
+            field: "isInternal",
             sortable: false,
             align: "center",
             class: "tight",
@@ -105,12 +105,12 @@
           },
           {
             title: this.$t('message.license'),
-            field: "component.license",
+            field: "license",
             sortable: false,
             formatter(value, row, index) {
-              if (Object.prototype.hasOwnProperty.call(row.component, "resolvedLicense")) {
-                let licenseurl = "../licenses/" + row.component.resolvedLicense.licenseId;
-                return "<a href=\"" + licenseurl + "\">" + xssFilters.inHTMLData(row.component.resolvedLicense.licenseId) + "</a>";
+              if (Object.prototype.hasOwnProperty.call(row, "resolvedLicense")) {
+                let licenseurl = "../licenses/" + row.resolvedLicense.licenseId;
+                return "<a href=\"" + licenseurl + "\">" + xssFilters.inHTMLData(row.resolvedLicense.licenseId) + "</a>";
               } else {
                 return xssFilters.inHTMLData(common.valueWithDefault(value, ""));
               }
@@ -118,7 +118,7 @@
           },
           {
             title: this.$t('message.risk_score'),
-            field: "component.lastInheritedRiskScore",
+            field: "lastInheritedRiskScore",
             sortable: true,
             class: "tight",
           },
@@ -167,7 +167,7 @@
             res.total = xhr.getResponseHeader("X-Total-Count");
             return res;
           },
-          url: `${this.$api.BASE_URL}/${this.$api.URL_DEPENDENCY}/project/${this.uuid}`
+          url: `${this.$api.BASE_URL}/${this.$api.URL_COMPONENT}/project/${this.uuid}`
         }
       };
     },
